Create datapower_audit.py
This is a simple python code built for fetching datapower servers audit log. Since datapower does not show audit log in file management we need to fetch it via xml request. This code assumes that number of log files rotation is 2 which will be audit-log and audit-log.1. For 3 or more rotation a different switch mechanism could be used after configuring audit log settings on datapower server.